Add files via upload
This is the data used for this project. The raw datafile "2004-16.csv" is downloaded from the website https://www.transtats.bts.gov/OT_Delay/OT_DelayCause1.asp. The raw file is then processed using  R and merged with the airport abrreviation and location files and thus obtained our dataset "2004-2016.csv" .